zombienet_configuration::shared
Module node

Structs

EnvVar

An environment variable with a name and a value. It can be constructed from a (&str, &str).

NodeConfig

A node configuration, with fine-grained configuration options.

NodeConfigBuilder

A node configuration builder, used to build a NodeConfig declaratively with fields validation.

Enums

Buildable

Initial