zombienet_configuration::shared
Module types

Structs

Chain

A chain name. It can be constructed for an &str, if it fails, it will returns a ConversionError.

ChainDefaultContext

Command

A command that will be executed natively (native provider) or in a container (podman/k8s). It can be constructed from an &str, if it fails, it will returns a ConversionError.

Image

A container image. It can be constructed from an &str including a combination of name, version, IPv4 or/and hostname, if it fails, it will returns a ConversionError.

U128

Custom type wrapping u128 to add custom Serialization/Deserialization logic because it’s not supported issue tracking the problem: https://github.com/toml-rs/toml/issues/540

ValidationContext

Enums

Arg

A CLI argument passed to an executed command, can be an option with an assigned value or a simple flag to enable/disable a feature. A flag arg can be constructed from a &str and a option arg can be constructed from a (&str, &str).

AssetLocation

A location for a locally or remotely stored asset. It can be constructed from an url::Url, a std::path::PathBuf or an &str.

Type Aliases

Duration

An alias for a duration in seconds.

ParaId

An alias for a parachain ID.

Port

An alias for a port.