Struct zombienet_configuration::ParachainConfigBuilder

pub struct ParachainConfigBuilder<S, C> { /* private fields */ }

A parachain configuration builder, used to build a ParachainConfig declaratively with fields validation.

Implementations

impl ParachainConfigBuilder<Initial, Bootstrap>

pub fn new( validation_context: Rc<RefCell<ValidationContext>>, ) -> ParachainConfigBuilder<Initial, Bootstrap>

Instantiate a new builder that can be used to build a ParachainConfig during the bootstrap phase.

impl ParachainConfigBuilder<WithId, Bootstrap>

pub fn with_registration_strategy(self, strategy: RegistrationStrategy) -> Self

Set the registration strategy for the parachain, could be Manual (no registered by zombienet) or automatic using an extrinsic or in genesis.

impl ParachainConfigBuilder<WithId, Running>

pub fn with_registration_strategy(self, strategy: RegistrationStrategy) -> Self

Set the registration strategy for the parachain, could be Manual (no registered by zombienet) or automatic Using an extrinsic. Genesis option is not allowed in Running context.

impl ParachainConfigBuilder<Initial, Running>

pub fn new_with_running( validation_context: Rc<RefCell<ValidationContext>>, ) -> ParachainConfigBuilder<Initial, Running>

Start a new builder in the context of a running network

impl<C: Context> ParachainConfigBuilder<Initial, C>

pub fn with_id(self, id: u32) -> ParachainConfigBuilder<WithId, C>

Set the parachain ID (should be unique).

impl<C: Context> ParachainConfigBuilder<WithId, C>

pub fn with_chain<T>(self, chain: T) -> Self
where T: TryInto<Chain>, T::Error: Error + Send + Sync + 'static,

Set the chain name (e.g. rococo-local). Use None, if you are running adder-collator or undying-collator).

pub fn onboard_as_parachain(self, choice: bool) -> Self

Set whether the parachain should be onboarded or stay a parathread. Default is true.

pub fn with_initial_balance(self, initial_balance: u128) -> Self

Set the initial balance of the parachain account.

pub fn with_default_command<T>(self, command: T) -> Self
where T: TryInto<Command>, T::Error: Error + Send + Sync + 'static,

Set the default command used for collators. Can be overridden.

pub fn with_default_image<T>(self, image: T) -> Self
where T: TryInto<Image>, T::Error: Error + Send + Sync + 'static,

Set the default container image used for collators. Can be overridden.

pub fn with_default_resources( self, f: impl FnOnce(ResourcesBuilder) -> ResourcesBuilder, ) -> Self

Set the default resources limits used for collators. Can be overridden.

pub fn with_default_db_snapshot( self, location: impl Into<AssetLocation>, ) -> Self

Set the default database snapshot location that will be used for state. Can be overridden.

pub fn with_default_args(self, args: Vec<Arg>) -> Self

Set the default arguments that will be used to execute the collator command. Can be overridden.

pub fn with_genesis_wasm_path(self, location: impl Into<AssetLocation>) -> Self

Set the location of a pre-existing genesis WASM runtime blob of the parachain.

pub fn with_genesis_wasm_generator<T>(self, command: T) -> Self
where T: TryInto<Command>, T::Error: Error + Send + Sync + 'static,

Set the generator command used to create the genesis WASM runtime blob of the parachain.

pub fn with_genesis_state_path(self, location: impl Into<AssetLocation>) -> Self

Set the location of a pre-existing genesis state of the parachain.

pub fn with_genesis_state_generator<T>(self, command: T) -> Self
where T: TryInto<Command>, T::Error: Error + Send + Sync + 'static,

Set the generator command used to create the genesis state of the parachain.

pub fn with_genesis_overrides(self, genesis_overrides: impl Into<Value>) -> Self

Set the genesis overrides as a JSON object.

pub fn with_chain_spec_path(self, location: impl Into<AssetLocation>) -> Self

Set the location of a pre-existing chain specification for the parachain.

pub fn with_chain_spec_command(self, cmd_template: impl Into<String>) -> Self

Set the chain-spec command template for the relay chain.

pub fn chain_spec_command_is_local(self, choice: bool) -> Self

Set if the chain-spec command needs to be run locally or not (false by default)

pub fn cumulus_based(self, choice: bool) -> Self

Set whether the parachain is based on cumulus (true in a majority of case, except adder or undying collators).

pub fn evm_based(self, choice: bool) -> Self

Set whether the parachain is evm based (e.g frontier /evm template)

pub fn with_bootnodes_addresses<T>(self, bootnodes_addresses: Vec<T>) -> Self
where T: TryInto<Multiaddr> + Display + Copy, T::Error: Error + Send + Sync + 'static,

Set the bootnodes addresses the collators will connect to.

pub fn with_collator( self, f: impl FnOnce(NodeConfigBuilder<Initial>) -> NodeConfigBuilder<Buildable>, ) -> ParachainConfigBuilder<WithAtLeastOneCollator, C>

Add a new collator using a nested NodeConfigBuilder.

impl<C: Context> ParachainConfigBuilder<WithAtLeastOneCollator, C>

pub fn with_collator( self, f: impl FnOnce(NodeConfigBuilder<Initial>) -> NodeConfigBuilder<Buildable>, ) -> Self

Add a new collator using a nested NodeConfigBuilder.

pub fn build(self) -> Result<ParachainConfig, Vec<Error>>

Seals the builder and returns a ParachainConfig if there are no validation errors, else returns errors.