zombienet_configuration::shared::node
Struct NodeConfigBuilder

pub struct NodeConfigBuilder<S> { /* private fields */ }

A node configuration builder, used to build a NodeConfig declaratively with fields validation.

Implementations

impl NodeConfigBuilder<Initial>

pub fn new( chain_context: ChainDefaultContext, validation_context: Rc<RefCell<ValidationContext>>, ) -> Self

pub fn with_name<T: Into<String> + Copy>( self, name: T, ) -> NodeConfigBuilder<Buildable>

Set the name of the node.

impl NodeConfigBuilder<Buildable>

pub fn with_command<T>(self, command: T) -> Self
where T: TryInto<Command>, T::Error: Error + Send + Sync + 'static,

Set the command that will be executed to launch the node. Override the default.

pub fn with_subcommand<T>(self, subcommand: T) -> Self
where T: TryInto<Command>, T::Error: Error + Send + Sync + 'static,

Set the subcommand that will be executed to launch the node.

pub fn with_image<T>(self, image: T) -> Self
where T: TryInto<Image>, T::Error: Error + Send + Sync + 'static,

Set the image that will be used for the node (only podman/k8s). Override the default.

pub fn with_args(self, args: Vec<Arg>) -> Self

Set the arguments that will be used when launching the node. Override the default.

pub fn validator(self, choice: bool) -> Self

Set whether the node is a validator.

pub fn invulnerable(self, choice: bool) -> Self

Set whether the node is invulnerable.

pub fn bootnode(self, choice: bool) -> Self

Set whether the node is a bootnode.

pub fn with_initial_balance(self, initial_balance: u128) -> Self

Set the node initial balance.

pub fn with_env(self, env: Vec<impl Into<EnvVar>>) -> Self

Set the node environment variables that will be used when launched. Override the default.

pub fn with_bootnodes_addresses<T>(self, bootnodes_addresses: Vec<T>) -> Self
where T: TryInto<Multiaddr> + Display + Copy, T::Error: Error + Send + Sync + 'static,

Set the bootnodes addresses that the node will try to connect to. Override the default.

pub fn with_resources( self, f: impl FnOnce(ResourcesBuilder) -> ResourcesBuilder, ) -> Self

Set the resources limits what will be used for the node (only podman/k8s). Override the default.

pub fn with_ws_port(self, ws_port: Port) -> Self

Set the websocket port that will be exposed. Uniqueness across config will be checked.

pub fn with_rpc_port(self, rpc_port: Port) -> Self

Set the RPC port that will be exposed. Uniqueness across config will be checked.

pub fn with_prometheus_port(self, prometheus_port: Port) -> Self

Set the Prometheus port that will be exposed for metrics. Uniqueness across config will be checked.

pub fn with_p2p_port(self, p2p_port: Port) -> Self

Set the P2P port that will be exposed. Uniqueness across config will be checked.

pub fn with_p2p_cert_hash(self, p2p_cert_hash: impl Into<String>) -> Self

Set the P2P cert hash that will be used as part of the multiaddress if and only if the multiaddress is set to use webrtc.

pub fn with_db_snapshot(self, location: impl Into<AssetLocation>) -> Self

Set the database snapshot that will be used to launch the node. Override the default.

pub fn build(self) -> Result<NodeConfig, (String, Vec<Error>)>

Seals the builder and returns a NodeConfig if there are no validation errors, else returns errors.